+++ src/Summoner/License.hs
@@ -4,12 +4,15 @@ module Summoner.License
        , customizeLicense
        , githubLicenseQueryNames
        , parseLicenseName
+       , fetchLicense
        ) where
 
 import Relude
 import Relude.Extra.Enum (inverseMap)
 
-import Data.Aeson (FromJSON (..), withObject, (.:))
+import Data.Aeson (FromJSON (..), decodeStrict, withObject, (.:))
+import Data.ByteString.Char8 (pack)
+import System.Process (readProcess)
 
 import qualified Data.Text as T
 import qualified Text.Show as TS
@@ -65,14 +68,21 @@ githubLicenseQueryNames = \case
 parseLicenseName :: Text -> Maybe LicenseName
 parseLicenseName = inverseMap show
 
-customizeLicense :: LicenseName -> Text -> Text -> Text -> Text
-customizeLicense l t nm year
-    | l `elem` [MIT, BSD2, BSD3] = updateLicenseText
-    | otherwise                  = t
+customizeLicense :: LicenseName -> License -> Text -> Text -> License
+customizeLicense l license@(License licenseText) nm year
+    | l `elem` [MIT, BSD2, BSD3] = License updateLicenseText
+    | otherwise                  = license
   where
     updateLicenseText =
-        let (beforeY, withY) = T.span (/= '[') t
+        let (beforeY, withY) = T.span (/= '[') licenseText
             afterY           = T.tail $ T.dropWhile (/= ']') withY
             (beforeN, withN) = T.span (/= '[') afterY
             afterN           = T.tail $ T.dropWhile (/= ']') withN
         in  beforeY <> year <> beforeN <> nm <> afterN
+
+fetchLicense :: LicenseName -> IO License
+fetchLicense name = do
+    let licenseLink = "https://api.github.com/licenses/" <> githubLicenseQueryNames name
+    licenseJson <- readProcess
+        "curl" [ toString licenseLink, "-H", "Accept: application/vnd.github.drax-preview+json"] ""
+    pure $ fromMaybe (error "Broken predefined license list") (decodeStrict $ pack licenseJson)